feat(decorators): Code split for better maintainability
