+++ src/interfaces/app/index.ts
@@ -1,3 +1,4 @@
 export * from "./controller-metadata.interface";
+export * from "./express-router-function-args.interface";
 export * from "./path-param-metadata.interface";
 export * from "./route-handler-metadata.interface";
